galaxys4gmtd: Use new bootanimation
While it is slightly cropped, its still recognizable.  The advantage to using it over the old one is that its not as static => less screen burn in

We could also create our own using the script present, but I do not want to have to maintain it.

Also, it does mean that when booting with gapps, we do run out of RAM, causing a reboot.